requirements.txt
streamlit
ultralytics
opencv-python-headless==4.7.0.72
numpy
Pillow
reportlab
